... and rename folder to match convention
